requirements.txt
alembic==1.8.1
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.5
cachelib==0.9.0
cachetools==5.5.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.5
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.8.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fqdn==1.5.1
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.72.0rc1
grpcio-status==1.71.0
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jiter==0.10.0
json5==0.9.14
jsonpointer==2.4
jsons==1.6.3
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
mysql-connector-python==8.0.31
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.4.2
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
openai==1.82.0
overrides==7.4.0
packaging==21.3
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pillow==11.1.0
platformdirs==2.5.2
playwright==1.50.0
pluggy==1.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.26.1
protobuf==5.29.4
psutil==5.9.5
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==22.3.5
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pyee==12.1.1
Pygments==2.16.1
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
pytest-order==1.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.7
pytz==2025.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.11.6
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9.1
safetensors==0.5.3
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
stack-data==0.6.2
sympy==1.14.0
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.21.1
tomli==2.0.1
torch==2.7.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.10.0
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
typish==1.9.3
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.16.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.2
widgetsnbextension==4.0.9
WTForms==3.0.1
zipp==3.9.0
